Add: server directory and change npm scripts for server
