+++ package.json
@@ -7,7 +7,7 @@
     "dev-server": "NODE_ENV=DEVELOPMENT webpack-dev-server --config webpack.dev.js",
     "dev": "NODE_ENV=DEVELOPMENT webpack --config webpack.dev.js",
     "build": "NODE_ENV=PRODUCTION webpack --config webpack.prod.js",
-    "start": "nodemon server.js"
+    "start": "nodemon server/server.js"
   },
   "keywords": [],
   "author": "Kee Wook Lee, James Flaws, Robert Garcia",
